update_CodeUp_C언어기초100제

+++ PracticeCodingTest/CodeUp_C언어기초100제/1030.cpp
@@ -6,7 +6,7 @@ int main()
 	ios::sync_with_stdio(false);
 	cin.tie(nullptr);
 
-	unsigned int n;
+	long long n;
 	cin >> n;
 	cout << n;
 }
